requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
beautifulsoup4==4.12.2
bootstrap4==0.1.0
Django==4.2.8
django-bootstrap-datepicker-plus==5.0.5
django-bootstrap4==23.2
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.1
pydantic==2.5.2
pydantic_core==2.14.5
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
